feat: split governance event details code into multiple files (#3255)
